--- casual/casual-inbound-handler-api/src/test/resources/META-INF/services/se.laz.casual.jca.inbound.handler.service.CasualServiceCallExtension
@@ -1,7 +1,0 @@
-#
-# Copyright (c) 2023, The casual project. All rights reserved.
-#
-# This software is licensed under the MIT license, https://opensource.org/licenses/MIT
-#
-se.laz.casual.jca.inbound.handler.test.TestCasualServiceCallExtensionLowestPriority
-se.laz.casual.jca.inbound.handler.test.TestCasualServiceCallExtensionHighestPriority
