+++ casual/casual-inbound-handler-api/src/test/resources/META-INF/services/se.laz.casual.jca.inbound.handler.service.CasualServiceHandlerExtension
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2023, The casual project. All rights reserved.
+#
+# This software is licensed under the MIT license, https://opensource.org/licenses/MIT
+#
+se.laz.casual.jca.inbound.handler.test.TestCasualServiceHandlerExtensionLowestPriority
+se.laz.casual.jca.inbound.handler.test.TestCasualServiceHandlerExtensionHighestPriority
